Move endpoint mounting in admin.go file
