+++ admin/admin.go
@@ -17,6 +17,25 @@ import (
 	"github.com/vechain/thor/v2/co"
 )
 
+func logLevelHandler(logLevel *slog.LevelVar) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		switch r.Method {
+		case http.MethodGet:
+			getLogLevelHandler(logLevel).ServeHTTP(w, r)
+		case http.MethodPost:
+			postLogLevelHandler(logLevel).ServeHTTP(w, r)
+		default:
+			writeError(w, http.StatusMethodNotAllowed, "method not allowed")
+		}
+	}
+}
+
+func HTTPHandler(logLevel *slog.LevelVar) http.Handler {
+	router := mux.NewRouter()
+	router.HandleFunc("/admin/loglevel", logLevelHandler(logLevel))
+	return handlers.CompressHandler(router)
+}
+
 func StartServer(addr string, logLevel *slog.LevelVar) (string, func(), error) {
 	listener, err := net.Listen("tcp", addr)
 	if err != nil {
