+++ admin/handlers.go
@@ -10,8 +10,6 @@ import (
 	"log/slog"
 	"net/http"
 
-	"github.com/gorilla/handlers"
-	"github.com/gorilla/mux"
 	"github.com/vechain/thor/v2/log"
 )
 
@@ -81,23 +79,4 @@ func postLogLevelHandler(logLevel *slog.LevelVar) http.HandlerFunc {
 		w.WriteHeader(http.StatusOK)
 		json.NewEncoder(w).Encode(response)
 	}
-}
-
-func logLevelHandler(logLevel *slog.LevelVar) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		switch r.Method {
-		case http.MethodGet:
-			getLogLevelHandler(logLevel).ServeHTTP(w, r)
-		case http.MethodPost:
-			postLogLevelHandler(logLevel).ServeHTTP(w, r)
-		default:
-			writeError(w, http.StatusMethodNotAllowed, "method not allowed")
-		}
-	}
-}
-
-func HTTPHandler(logLevel *slog.LevelVar) http.Handler {
-	router := mux.NewRouter()
-	router.HandleFunc("/admin/loglevel", logLevelHandler(logLevel))
-	return handlers.CompressHandler(router)
 }
